library r2jags to ensure dependencies are loaded
i could make the whoel trawlDiversity package depend on R2jags, but what if I just wanted to use Stan, and not install R2jags? Then I'd have to install it just to not use it. I think this is the easiest way. Problem with just namespace is that I think inside some of the R2jags functions package::fun() convention is ignored, and so some of the functions aren't correctly found