if reshape is loaded, reshape2::melt() breaks
this is because reshape2::melt() will actually refer to reshape::melt.table or reshape::melt.data.frame. This doesn't happen if the function is run in a clean R session, but if reshape happens to be loaded after reshape2, then this bites, hard. The difference between reshape and reshape2 is that the value.name argument of reshape::melt does not properly reassign the column name.